requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
python-socketio>=5.11.1
python-engineio>=4.9.1
pandas>=2.0.0
openpyxl>=3.1.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
selenium>=4.20.0
webdriver-manager>=4.0.0
requests>=2.31.0
pyngrok>=7.0.0
gunicorn>=21.2.0
eventlet>=0.36.0
